requirements.txt
absl-py==2.0.0
asgiref==3.7.2
astor==0.8.1
cycler==0.11.0
Django==3.2.23
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.8.0
importlib-metadata==6.7.0
joblib==1.3.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
